requirements.txt
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
Pillow==7.0.0
